comment error handle

+++ src/main/java/helper/recruit/community/dto/QuestionDTO.java
@@ -11,7 +11,7 @@ public class QuestionDTO {
     private String tag;
     private Long gmtCreate;
     private Long gmtModified;
-    private Integer creator;  // connect to user ID in USERS Table (database)
+    private Long creator;  // connect to user ID in USERS Table (database)
     private Integer viewCount;
     private Integer commentCount;
     private Integer likeCount;
